+++ source/add_extra_stereo_audio/plugin.py
@@ -23,6 +23,7 @@
 """
 import logging
 import os
+import iso639
 
 from unmanic.libs.unplugins.settings import PluginSettings
 
@@ -40,6 +41,7 @@ class Settings(PluginSettings):
         "use_libfdk_aac":         False,
         "remove_original_multichannel_audio": False,
         "make_xtra_stereo_default": True,
+        "move_to_first"           : True,
     }
 
     def __init__(self, *args, **kwargs):
@@ -60,9 +62,11 @@ class Settings(PluginSettings):
             "remove_original_multichannel_audio":               {
                 "label": "check if you want to remove the original multichannel audio stream",
             },
-
             "make_xtra_stereo_default":      {
-                "label": "check if you want the new stereo audio to be the default audio upon playing"
+                "label": "check if you want the new stereo audio to be the default audio upon playing",
+            },
+            "move_to_first":                 {
+                "label": "check if you want to move the stereo audio to first audio stream"
             }
         }
 
@@ -150,7 +154,7 @@ def on_library_management_file_test(data):
         else:
             logger.debug("Audio stream '{}' is '{}' and has '{}' channels and is encoded with '{}' - convert stream".format(stream, stream_language, channels, codec_name))
     else:
-        data['add_file_to_pending_tasks'] = False
+#        data['add_file_to_pending_tasks'] = False
         logger.debug("do not add file '{}' to task list - no matching streams".format(abspath))
 
     return data
@@ -209,25 +213,58 @@ def on_worker_process(data):
     if settings.get_setting('use_libfdk_aac'): encoder = 'libfdk_aac'
     remove_original = settings.get_setting('remove_original_multichannel_audio')
     set_default_audio_to_new_stream = settings.get_setting('make_xtra_stereo_default')
+    move_to_first = settings.get_setting('move_to_first')
 
     stream = stream_to_stereo_encode(stream_language, channels, codec_name, probe_streams)
     if stream >= 0:
 
+        # initialize ffmpeg_args
+        ffmpeg_args = ['-hide_banner', '-loglevel', 'info', '-i', str(abspath), '-max_muxing_queue_size', '9999']
+
         # Get generated ffmpeg args
-        if not remove_original:
-            ffmpeg_args = ['-hide_banner', '-loglevel', 'info', '-i', str(abspath), '-max_muxing_queue_size', '9999', '-map', '0', '-c', 'copy', '-map', '0:a:'+str(stream), '-c:a:'+str(new_audio_stream), encoder, '-ac', '2', '-b:a:'+str(new_audio_stream), '128k']
-        else:
-            ffmpeg_args = ['-hide_banner', '-loglevel', 'info', '-i', str(abspath), '-max_muxing_queue_size', '9999', '-map', '0:s?', '-c:s', 'copy', '-map', '0:d?', '-c:d', 'copy', '-map', '0:t?', '-c:t', 'copy', '-map', '0:v', '-c:v', 'copy']
+        if not remove_original and not move_to_first:
+            ffmpeg_args += ['-map', '0', '-c', 'copy', '-map', '0:a:'+str(stream), '-c:a:'+str(new_audio_stream), encoder, '-ac', '2', '-b:a:'+str(new_audio_stream), '128k']
+        elif remove_original and not move_to_first:
+            ffmpeg_args += ['-map', '0:s?', '-c:s', 'copy', '-map', '0:d?', '-c:d', 'copy', '-map', '0:t?', '-c:t', 'copy', '-map', '0:v', '-c:v', 'copy']
             for astream in range(0, total_audio_streams+1):
                 if astream ==  stream:
                     ffmpeg_args += ['-map', '0:a:'+str(stream), '-c:a:'+str(stream), encoder, '-ac', '2', '-b:a:'+str(stream), '128k']
                     new_audio_stream = astream
                 else:
                     ffmpeg_args += ['-map', '0:a:'+str(astream), '-c:a:'+str(astream), 'copy']
+        elif move_to_first:
+            new_audio_stream = 0
+            ffmpeg_args += ['-map', '0:s?', '-c:s', 'copy', '-map', '0:d?', '-c:d', 'copy', '-map', '0:t?', '-c:t', 'copy', '-map', '0:v', '-c:v', 'copy']
+
+            # place stream_to_stereo_encode as 1st audio:
+            ffmpeg_args += ['-map', '0:a:'+str(stream), '-c:a:0', encoder, '-ac', '2', '-b:a:0', '128k']
+
+            # iterate over remaining streams and copy unless it's to be removed
+            skip_original = 0
+            for astream in range(0, total_audio_streams+1):
+                if astream == stream and remove_original:
+                    skip_original = 1
+                else:
+                    ffmpeg_args += ['-map', '0:a:'+str(astream), '-c:a:'+str(astream + 1 - skip_original), 'copy']
+
         if set_default_audio_to_new_stream:
-            ffmpeg_args += ['-disposition:a', '-default', '-disposition:a:'+str(new_audio_stream), 'default', '-y', str(outpath)]
-        else:
-            ffmpeg_args += ['-y', str(outpath)]
+            ffmpeg_args += ['-disposition:a', '-default', '-disposition:a:'+str(new_audio_stream), 'default']
+
+        logger.debug("stream_language: '{}'".format(stream_language))
+
+        try:
+            if len(stream_language) == 2:
+                lang = iso639.Language.from_part1(stream_language)
+            elif len(stream_language) == 3:
+                lang = iso639.Language.from_part3(stream_language)
+        except iso639.language.LanguageNotFoundError:
+            logger.info("iso 639 exception")
+            lang = stream_language
+
+        if lang != stream_language: lang = lang.name
+
+        logger.debug("lang: '{}'".format(lang))
+        ffmpeg_args += ['-metadata:s:a:' + str(new_audio_stream), 'title=' + str(lang) + ' (' +str(encoder).upper() + ' Stereo)', '-y', str(outpath)]
 
         logger.debug("ffmpeg args: '{}'".format(ffmpeg_args))